+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 
 ### Added
 
+- `FST.docstr` property
 - automatic coercion of nodes put as `AST` or `FST` to compatible types
 - replacing all handlers in a `Try` or `TryStar` with the other kind will change the try `AST` to the other kind
 - prescribed slicing for `Compare` (`left` + `comparators` as a single body)